basis for puzzle
would like to have a more exciting puzzle, but as is, should offer a simple solution to connect other puzzles/final door code.
You can input a number in the inpufield and test the input with the green "button".
The idea is that you start in Deut -> Protan -> Tritan -> normal vision
the middle cube might be too hard to see at any point.
I am thinking about punishment for trying to guess too many times.